Split inclusionary_zoning.rs (744 lines) into subdirectory modules
Replace the single-file module with a subdirectory containing:
- config.rs (114 lines): constants, DistrictInclusionaryConfig, InclusionaryZoningState
- helpers.rs (69 lines): pure calculation functions (FAR bonus, affordable units, etc.)
- systems.rs (76 lines): ECS system and Saveable implementation
- tests.rs (452 lines): all unit tests
- mod.rs (50 lines): module declarations, re-exports, and plugin

All public items are re-exported from mod.rs so external code is unaffected.

Part of #1389
