+++ crates/simulation/src/inclusionary_zoning/tests.rs
@@ -1,301 +1,9 @@
-//! Inclusionary Zoning Requirements (ZONE-010).
-//!
-//! Implements inclusionary zoning as a district policy requiring new residential
-//! developments to reserve a percentage of units as affordable housing.
-//!
-//! Features:
-//! - **District policy toggle**: Inclusionary Zoning can be enabled per player-defined district
-//! - **Affordable unit percentage**: 10-20% of residential units reserved as affordable housing
-//! - **FAR bonus**: +10-20% Floor Area Ratio bonus to offset affordable unit costs
-//! - **Affordable units**: House lower-income citizens who would otherwise be priced out
-//! - **Profitability impact**: Affects building profitability and construction rate
-//!
-//! The system tracks which districts have inclusionary zoning enabled and computes
-//! per-district effects (effective capacity reduction, FAR bonus, affordable unit counts).
-
-use bevy::prelude::*;
-use bitcode::{Decode, Encode};
-use std::collections::HashMap;
-
-use crate::buildings::Building;
-use crate::districts::DistrictMap;
-use crate::SlowTickTimer;
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// Minimum affordable housing percentage (10%).
-pub const MIN_AFFORDABLE_PERCENTAGE: f32 = 0.10;
-
-/// Maximum affordable housing percentage (20%).
-pub const MAX_AFFORDABLE_PERCENTAGE: f32 = 0.20;
-
-/// Default affordable housing percentage when policy is first enabled (15%).
-pub const DEFAULT_AFFORDABLE_PERCENTAGE: f32 = 0.15;
-
-/// Minimum FAR bonus granted to offset affordable unit cost (10%).
-pub const MIN_FAR_BONUS: f32 = 0.10;
-
-/// Maximum FAR bonus granted to offset affordable unit cost (20%).
-pub const MAX_FAR_BONUS: f32 = 0.20;
-
-/// Construction rate penalty multiplier when inclusionary zoning is active.
-/// Developers build slightly slower due to reduced profitability.
-pub const CONSTRUCTION_RATE_PENALTY: f32 = 0.90;
-
-/// Monthly administrative cost per district with inclusionary zoning.
-pub const MONTHLY_ADMIN_COST_PER_DISTRICT: f64 = 8_000.0;
-
-// =============================================================================
-// Per-district configuration
-// =============================================================================
-
-/// Per-district inclusionary zoning configuration.
-#[derive(Debug, Clone, Encode, Decode)]
-pub struct DistrictInclusionaryConfig {
-    /// Whether inclusionary zoning is enabled for this district.
-    pub enabled: bool,
-    /// Percentage of units that must be affordable (0.10 - 0.20).
-    pub affordable_percentage: f32,
-}
-
-impl Default for DistrictInclusionaryConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            affordable_percentage: DEFAULT_AFFORDABLE_PERCENTAGE,
-        }
-    }
-}
-
-// =============================================================================
-// Resource: Inclusionary Zoning State
-// =============================================================================
-
-/// Tracks inclusionary zoning policy state across all player-defined districts.
-#[derive(Resource, Debug, Clone, Default, Encode, Decode)]
-pub struct InclusionaryZoningState {
-    /// Per-district inclusionary zoning configurations.
-    /// Key is the district index in the DistrictMap.
-    pub district_configs: HashMap<usize, DistrictInclusionaryConfig>,
-    /// Total affordable units across all districts (computed).
-    pub total_affordable_units: u32,
-    /// Total residential units in affected districts (computed).
-    pub total_affected_units: u32,
-    /// Total monthly admin cost (computed).
-    pub total_monthly_cost: f64,
-}
-
-impl InclusionaryZoningState {
-    /// Enable inclusionary zoning for a district with the default percentage.
-    pub fn enable(&mut self, district_idx: usize) {
-        let config = self.district_configs.entry(district_idx).or_default();
-        config.enabled = true;
-    }
-
-    /// Disable inclusionary zoning for a district.
-    pub fn disable(&mut self, district_idx: usize) {
-        if let Some(config) = self.district_configs.get_mut(&district_idx) {
-            config.enabled = false;
-        }
-    }
-
-    /// Check if inclusionary zoning is enabled for a district.
-    pub fn is_enabled(&self, district_idx: usize) -> bool {
-        self.district_configs
-            .get(&district_idx)
-            .is_some_and(|c| c.enabled)
-    }
-
-    /// Set the affordable percentage for a district (clamped to valid range).
-    pub fn set_affordable_percentage(&mut self, district_idx: usize, pct: f32) {
-        let clamped = pct.clamp(MIN_AFFORDABLE_PERCENTAGE, MAX_AFFORDABLE_PERCENTAGE);
-        let config = self.district_configs.entry(district_idx).or_default();
-        config.affordable_percentage = clamped;
-    }
-
-    /// Get the affordable percentage for a district (returns 0.0 if not enabled).
-    pub fn affordable_percentage(&self, district_idx: usize) -> f32 {
-        self.district_configs
-            .get(&district_idx)
-            .filter(|c| c.enabled)
-            .map(|c| c.affordable_percentage)
-            .unwrap_or(0.0)
-    }
-
-    /// Get the number of enabled districts.
-    pub fn enabled_district_count(&self) -> usize {
-        self.district_configs.values().filter(|c| c.enabled).count()
-    }
-}
-
-// =============================================================================
-// Pure helper functions (testable without ECS)
-// =============================================================================
-
-/// Calculate the FAR bonus for a given affordable percentage.
-/// The bonus scales linearly from MIN_FAR_BONUS at MIN_AFFORDABLE_PERCENTAGE
-/// to MAX_FAR_BONUS at MAX_AFFORDABLE_PERCENTAGE.
-pub fn calculate_far_bonus(affordable_pct: f32) -> f32 {
-    if affordable_pct <= 0.0 {
-        return 0.0;
-    }
-    let clamped = affordable_pct.clamp(MIN_AFFORDABLE_PERCENTAGE, MAX_AFFORDABLE_PERCENTAGE);
-    let t = (clamped - MIN_AFFORDABLE_PERCENTAGE)
-        / (MAX_AFFORDABLE_PERCENTAGE - MIN_AFFORDABLE_PERCENTAGE);
-    MIN_FAR_BONUS + t * (MAX_FAR_BONUS - MIN_FAR_BONUS)
-}
-
-/// Calculate the number of affordable units for a building given its capacity
-/// and the district's affordable percentage.
-pub fn calculate_affordable_units(capacity: u32, affordable_pct: f32) -> u32 {
-    if affordable_pct <= 0.0 {
-        return 0;
-    }
-    // Round to nearest; guarantee at least 1 unit if building has capacity and policy is active
-    let raw = (capacity as f32 * affordable_pct).round() as u32;
-    let raw = if capacity > 0 { raw.max(1) } else { raw };
-    raw.min(capacity)
-}
-
-/// Calculate the effective (market-rate) capacity after removing affordable units.
-pub fn calculate_effective_capacity(capacity: u32, affordable_pct: f32) -> u32 {
-    let affordable = calculate_affordable_units(capacity, affordable_pct);
-    capacity.saturating_sub(affordable)
-}
-
-/// Calculate the monthly admin cost for the given number of enabled districts.
-pub fn calculate_monthly_admin_cost(enabled_count: usize) -> f64 {
-    enabled_count as f64 * MONTHLY_ADMIN_COST_PER_DISTRICT
-}
-
-/// Check if a cell is in a district with inclusionary zoning enabled.
-pub fn is_cell_in_inclusionary_district(
-    x: usize,
-    y: usize,
-    state: &InclusionaryZoningState,
-    district_map: &DistrictMap,
-) -> bool {
-    district_map
-        .get_district_index_at(x, y)
-        .is_some_and(|di| state.is_enabled(di))
-}
-
-/// Get the affordable percentage for the cell's district, or 0.0 if not in
-/// an inclusionary zoning district.
-pub fn affordable_percentage_for_cell(
-    x: usize,
-    y: usize,
-    state: &InclusionaryZoningState,
-    district_map: &DistrictMap,
-) -> f32 {
-    district_map
-        .get_district_index_at(x, y)
-        .map(|di| state.affordable_percentage(di))
-        .unwrap_or(0.0)
-}
-
-// =============================================================================
-// Systems
-// =============================================================================
-
-/// System: update inclusionary zoning computed effects every slow tick.
-///
-/// Iterates all residential buildings, checks if they are in a district
-/// with inclusionary zoning enabled, and aggregates total affordable/affected
-/// unit counts and admin costs.
-pub fn update_inclusionary_zoning(
-    timer: Res<SlowTickTimer>,
-    mut state: ResMut<InclusionaryZoningState>,
-    district_map: Res<DistrictMap>,
-    buildings: Query<&Building>,
-) {
-    if !timer.should_run() {
-        return;
-    }
-
-    let mut total_affordable = 0u32;
-    let mut total_affected = 0u32;
-
-    for building in &buildings {
-        if !building.zone_type.is_residential() && !building.zone_type.is_mixed_use() {
-            continue;
-        }
-
-        let di = district_map.get_district_index_at(building.grid_x, building.grid_y);
-        let affordable_pct = di
-            .map(|idx| state.affordable_percentage(idx))
-            .unwrap_or(0.0);
-
-        if affordable_pct > 0.0 {
-            let res_capacity = if building.zone_type.is_mixed_use() {
-                let (_, res_cap) =
-                    crate::buildings::MixedUseBuilding::capacities_for_level(building.level);
-                res_cap
-            } else {
-                building.capacity
-            };
-            total_affected += res_capacity;
-            total_affordable += calculate_affordable_units(res_capacity, affordable_pct);
-        }
-    }
-
-    state.total_affordable_units = total_affordable;
-    state.total_affected_units = total_affected;
-    state.total_monthly_cost = calculate_monthly_admin_cost(state.enabled_district_count());
-}
-
-// =============================================================================
-// Saveable implementation
-// =============================================================================
-
-impl crate::Saveable for InclusionaryZoningState {
-    const SAVE_KEY: &'static str = "inclusionary_zoning";
-
-    fn save_to_bytes(&self) -> Option<Vec<u8>> {
-        // Skip saving if no districts have ever been configured
-        if self.district_configs.is_empty() {
-            return None;
-        }
-        Some(bitcode::encode(self))
-    }
-
-    fn load_from_bytes(bytes: &[u8]) -> Self {
-        crate::decode_or_warn(Self::SAVE_KEY, bytes)
-    }
-}
-
-// =============================================================================
-// Plugin
-// =============================================================================
-
-pub struct InclusionaryZoningPlugin;
-
-impl Plugin for InclusionaryZoningPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<InclusionaryZoningState>().add_systems(
-            FixedUpdate,
-            update_inclusionary_zoning
-                .after(crate::buildings::progress_construction)
-                .in_set(crate::SimulationSet::Simulation),
-        );
-
-        // Register for save/load via the SaveableRegistry.
-        app.init_resource::<crate::SaveableRegistry>();
-        app.world_mut()
-            .resource_mut::<crate::SaveableRegistry>()
-            .register::<InclusionaryZoningState>();
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
+//! Tests for inclusionary zoning module.
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::districts::DistrictMap;
+    use crate::inclusionary_zoning::*;
 
     // -------------------------------------------------------------------------
     // Default state tests
